+++ lib/tests/web/roundtrip.test.ts
@@ -1,8 +1,6 @@
 import { expect } from '@esm-bundle/chai';
 import { type AuthProvider, HttpRequest, withHeaders } from '../../src/auth/auth.js';
 
-import { NanoTDFClient } from '../../src/nanoclients.js';
-import NanoTDF from '../../src/nanotdf/NanoTDF.js';
 import { OpenTDF } from '../../src/opentdf.js';
 import { fromString } from '../../src/seekable.js';
 
@@ -44,14 +42,14 @@ describe('Local roundtrip Tests', () => {
     });
     const cipherTextArray = new Uint8Array(await new Response(cipherTextStream).arrayBuffer());
 
-    const nanotdfParsed = await client.read({
+    const ztdfParsed = await client.read({
       source: { type: 'buffer', location: cipherTextArray },
     });
 
-    const metadata = (await nanotdfParsed.metadata) as never;
+    const metadata = (await ztdfParsed.metadata) as never;
     expect(metadata['hello']['kind']['value']).to.equal('world');
 
-    const actual = await new Response(nanotdfParsed).arrayBuffer();
+    const actual = await new Response(ztdfParsed).arrayBuffer();
     expect(new TextDecoder().decode(actual)).to.be.equal('hello world');
   });
   it(`ztdf roundtrip string with platformUrl`, async () => {
@@ -75,14 +73,14 @@ describe('Local roundtrip Tests', () => {
     });
     const cipherTextArray = new Uint8Array(await new Response(cipherTextStream).arrayBuffer());
 
-    const nanotdfParsed = await client.read({
+    const ztdfParsed = await client.read({
       source: { type: 'buffer', location: cipherTextArray },
     });
 
-    const metadata = (await nanotdfParsed.metadata) as never;
+    const metadata = (await ztdfParsed.metadata) as never;
     expect(metadata['hello']['kind']['value']).to.equal('world');
 
-    const actual = await new Response(nanotdfParsed).arrayBuffer();
+    const actual = await new Response(ztdfParsed).arrayBuffer();
     const decoded = new TextDecoder().decode(actual);
     expect(decoded).to.be.equal('hello world');
   });
@@ -136,52 +134,15 @@ describe('Local roundtrip Tests', () => {
     });
     const cipherTextArray = new Uint8Array(await new Response(cipherTextStream).arrayBuffer());
 
-    const nanotdfParsed = await client.read({
+    const ztdfParsed = await client.read({
       source: { type: 'buffer', location: cipherTextArray },
     });
 
-    const metadata = (await nanotdfParsed.metadata) as never;
+    const metadata = (await ztdfParsed.metadata) as never;
     expect(metadata['hello']['kind']['value']).to.equal('world');
 
-    const actual = await new Response(nanotdfParsed).arrayBuffer();
+    const actual = await new Response(ztdfParsed).arrayBuffer();
     const decoded = new TextDecoder().decode(actual);
     expect(decoded).to.be.equal('hello world');
   });
-
-  for (const ecdsaBinding of [false, true]) {
-    const bindingType = ecdsaBinding ? 'ecdsa' : 'gmac';
-    it(`nano roundtrip string (${bindingType} policy binding)`, async () => {
-      const client = new OpenTDF({
-        authProvider,
-        defaultReadOptions: {
-          allowedKASEndpoints: [kasEndpoint],
-        },
-      });
-      const cipherText = await client.createNanoTDF({
-        bindingType,
-        defaultKASEndpoint: kasEndpoint,
-        source: { type: 'chunker', location: fromString('hello world') },
-      });
-      const nanotdfParsed = await client.read({
-        source: { type: 'stream', location: cipherText },
-      });
-      expect(nanotdfParsed.header?.kas?.url).to.equal(kasEndpoint);
-      expect(nanotdfParsed.header?.kas?.identifier).to.equal('e1');
-
-      const actual = await new Response(nanotdfParsed).arrayBuffer();
-      expect(new TextDecoder().decode(actual)).to.be.equal('hello world');
-    });
-    it(`roundtrip string (${bindingType} policy binding, deprecated API)`, async () => {
-      const client = new NanoTDFClient({ authProvider, kasEndpoint, platformUrl });
-      const cipherText = await client.encrypt('hello world', { ecdsaBinding });
-      const client2 = new NanoTDFClient({ authProvider, kasEndpoint, platformUrl });
-      const nanotdfParsed = NanoTDF.from(cipherText);
-
-      expect(nanotdfParsed.header.kas.url).to.equal(kasEndpoint);
-      expect(nanotdfParsed.header.kas.identifier).to.equal('e1');
-
-      const actual = await client2.decrypt(cipherText);
-      expect(new TextDecoder().decode(actual)).to.be.equal('hello world');
-    });
-  }
 });